Reports foe customer added

+++ src/main/java/ua/heatloss/dao/UserDao.java
@@ -6,5 +6,5 @@ import ua.heatloss.domain.user.User;
 public interface UserDao extends CrudDao<User>{
     User getUserByEmail(String email);
 
-    int countUsersWithEmail(String email);
+    long countUsersWithEmail(String email);
 }
